requirements.txt
absl-py==1.3.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-vision==3.1.4
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.0
jupyter-events==0.5.0
jupyter_server==2.0.0
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.1.2.30
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==1.13.0
torchaudio==0.4.0a0+ba6d67b
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.3
zipp==3.11.0
